Optimize codebase
